refactor(remediate): split the executor out of cli.ts (module size)
The #273 wrap + seams pushed cli.ts over the large-file threshold and the
self-guardrail blocked it — the fix is the standing pattern (plan-cli /
configured-loop precedent), not an allowlist: the executor layer
(executeTask, the attempt record, the landing wrap, the seams) moves to
src/remediate/execute.ts and the CLI re-exports it, so consumers keep one
import surface.
